AS: move grpc CoCo-AS as a bin
CoCoAS is actually a binary rather than a lib. The commit would help to
move it under attestation service crate for better code organization.
