+++ attestation-service/.dockerignore
@@ -1,4 +1,4 @@
 target
 
-Dockerfile.as
+Dockerfile.as*
 Dockerfile.rvps
