+++ attestation-service/Dockerfile.as-grpc
@@ -22,7 +22,7 @@ RUN curl -L https://download.01.org/intel-sgx/sgx_repo/ubuntu/intel-sgx-deb.key 
     apt-get update && apt-get install -y libtdx-attest-dev libsgx-dcap-quote-verify-dev
 
 # Build and Install gRPC attestation-service
-RUN cargo install --locked --path attestation-service/bin/grpc-as
+RUN cargo install --path attestation-service/attestation-service --bin grpc-as --features grpc-bin --locked
 
 
 FROM ubuntu:22.04
