+++ attestation-service/README.md
@@ -32,7 +32,7 @@ attestation-service = { git = "https://github.com/confidential-containers/kbs", 
 
 This project provides the Attestation Service binary program that can be run as an independent server:
 
-- [`grpc-as`](bin/grpc-as/): Provide AS APIs based on gRPC protocol.
+- [`grpc-as`](docs/grpc-as.md): Provide AS APIs based on gRPC protocol.
 
 # Usage
