--- attestation-service/bin/grpc-as/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "grpc-as"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-anyhow.workspace = true
-async-trait.workspace = true
-attestation-service = { path = "../../attestation-service", features = ["rvps-grpc"] }
-base64.workspace = true
-clap.workspace = true
-env_logger.workspace = true
-log.workspace = true
-prost.workspace = true
-serde_json.workspace = true
-shadow-rs.workspace = true
-tokio.workspace = true
-tonic.workspace = true
-
-[build-dependencies]
-shadow-rs.workspace = true
-tonic-build.workspace = true
